+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivity.kt
@@ -12,24 +12,34 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.appcompat.widget.Toolbar
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.core.view.isVisible
 import androidx.core.view.updatePadding
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commitNow
+import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.LiveDataReactiveStreams
 import androidx.lifecycle.Observer
+import androidx.lifecycle.flowWithLifecycle
+import androidx.lifecycle.lifecycleScope
 import androidx.transition.Slide
 import au.com.shiftyjelly.pocketcasts.R
 import au.com.shiftyjelly.pocketcasts.account.AccountActivity
 import au.com.shiftyjelly.pocketcasts.account.PromoCodeUpgradedFragment
+import au.com.shiftyjelly.pocketcasts.account.onboarding.OnboardingActivity
+import au.com.shiftyjelly.pocketcasts.account.onboarding.OnboardingActivityContract
+import au.com.shiftyjelly.pocketcasts.account.onboarding.OnboardingActivityContract.OnboardingFinish
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.analytics.FirebaseAnalyticsTracker
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.databinding.ActivityMainBinding
 import au.com.shiftyjelly.pocketcasts.discover.view.DiscoverFragment
-import au.com.shiftyjelly.pocketcasts.endofyear.StoriesFragment
+import au.com.shiftyjelly.pocketcasts.endofyear.StoriesPage
 import au.com.shiftyjelly.pocketcasts.endofyear.views.EndOfYearLaunchBottomSheet
 import au.com.shiftyjelly.pocketcasts.filters.FiltersFragment
 import au.com.shiftyjelly.pocketcasts.localization.helper.LocaliseHelper
@@ -104,6 +114,15 @@ import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.addTo
 import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.withContext
 import timber.log.Timber
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
@@ -114,6 +133,7 @@ import au.com.shiftyjelly.pocketcasts.views.R as VR
 import com.google.android.material.R as MR
 
 private const val SAVEDSTATE_PLAYER_OPEN = "player_open"
+private const val SAVEDSTATE_MINIPLAYER_SHOWN = "miniplayer_shown"
 
 @AndroidEntryPoint
 class MainActivity :
@@ -174,13 +194,25 @@ class MainActivity :
         super.onCreate(savedInstanceState)
         theme.setupThemeForConfig(this, resources.configuration)
 
+        // TODO check settings to determine if onboarding has already been completed
+        if (BuildConfig.ONBOARDING_ENABLED && !settings.isLoggedIn()) {
+            openOnboardingFlow()
+        }
+
         binding = ActivityMainBinding.inflate(layoutInflater)
         val view = binding.root
         setContentView(view)
 
-        if (BuildConfig.END_OF_YEAR_ENABLED && settings.getEndOfYearShowBadge2022()) {
-            binding.bottomNavigation.getOrCreateBadge(VR.id.navigation_profile)
-        }
+        viewModel.isEndOfYearStoriesEligible()
+            .flowWithLifecycle(lifecycle, Lifecycle.State.STARTED)
+            .onEach { isEligible ->
+                if (isEligible) {
+                    setupEndOfYearLaunchBottomSheet()
+                    if (settings.getEndOfYearShowBadge2022()) {
+                        binding.bottomNavigation.getOrCreateBadge(VR.id.navigation_profile)
+                    }
+                }
+            }.launchIn(lifecycleScope)
 
         var selectedTab = settings.selectedTab()
         val tabs = mapOf(
@@ -213,7 +245,10 @@ class MainActivity :
             activity = this
         )
 
-        setupPlayerViews()
+        val showMiniPlayerImmediately = savedInstanceState?.getBoolean(SAVEDSTATE_MINIPLAYER_SHOWN, false) ?: false
+        binding.playerBottomSheet.isVisible = showMiniPlayerImmediately
+
+        setupPlayerViews(showMiniPlayerImmediately)
 
         if (savedInstanceState != null) {
             val videoComingToPortrait =
@@ -255,10 +290,22 @@ class MainActivity :
         handleIntent(intent, savedInstanceState)
 
         updateSystemColors()
+    }
 
-        if (BuildConfig.END_OF_YEAR_ENABLED) {
-            setupEndOfYearLaunchBottomSheet()
-        }
+    private fun openOnboardingFlow() {
+        registerForActivityResult(OnboardingActivityContract()) { result ->
+            when (result) {
+                OnboardingFinish.CompletedOnboarding -> {
+                    // TODO persist that onboarding has been completed
+                }
+                OnboardingFinish.AbortedOnboarding -> {
+                    finish()
+                }
+                null -> {
+                    Timber.e("Unexpected null result from onboarding activity")
+                }
+            }
+        }.launch(Intent(this, OnboardingActivity::class.java))
     }
 
     override fun onStart() {
@@ -307,6 +354,7 @@ class MainActivity :
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
         outState.putBoolean(SAVEDSTATE_PLAYER_OPEN, binding.playerBottomSheet.isPlayerOpen)
+        outState.putBoolean(SAVEDSTATE_MINIPLAYER_SHOWN, binding.playerBottomSheet.isShown)
     }
 
     override fun overrideNextRefreshTimer() {
@@ -458,11 +506,17 @@ class MainActivity :
 
     private fun setupEndOfYearLaunchBottomSheet() {
         binding.modalBottomSheet.setContent {
-            AppTheme(themeType = theme.activeTheme) {
+            var showDialog by rememberSaveable { mutableStateOf(false) }
+            if (showDialog) {
+                StoriesPage(
+                    theme = theme,
+                    onCloseClicked = { showDialog = false },
+                )
+            }
+            AppTheme(theme.activeTheme) {
                 EndOfYearLaunchBottomSheet(
                     onClick = {
-                        StoriesFragment.newInstance()
-                            .show(supportFragmentManager, "stories_dialog")
+                        showDialog = true
                     }
                 )
             }
@@ -471,7 +525,7 @@ class MainActivity :
 
     @OptIn(DelicateCoroutinesApi::class)
     @Suppress("DEPRECATION")
-    private fun setupPlayerViews() {
+    private fun setupPlayerViews(showMiniPlayerImmediately: Boolean) {
         binding.playerBottomSheet.listener = this
 
         viewModel.playbackState.observe(this) { state ->
@@ -494,6 +548,10 @@ class MainActivity :
                 updateNavAndStatusColors(true, state.podcast)
             }
 
+            if (viewModel.lastPlaybackState != null && (viewModel.lastPlaybackState?.episodeUuid != state.episodeUuid || (viewModel.lastPlaybackState?.isPlaying == false && state.isPlaying)) && settings.openPlayerAutomatically()) {
+                binding.playerBottomSheet.openPlayer()
+            }
+
             updatePlaybackState(state)
 
             viewModel.lastPlaybackState = state
@@ -509,7 +567,11 @@ class MainActivity :
                 .toFlowable(BackpressureStrategy.LATEST)
         observeUpNext = LiveDataReactiveStreams.fromPublisher(upNextQueueObservable)
         observeUpNext.observe(this) { upNext ->
-            binding.playerBottomSheet.setUpNext(upNext, theme)
+            binding.playerBottomSheet.setUpNext(
+                upNext = upNext,
+                theme = theme,
+                shouldAnimateOnAttach = !showMiniPlayerImmediately
+            )
         }
 
         viewModel.signInState.observe(this) { signinState ->