+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/helper/ListenedCategory.kt
@@ -1,0 +1,13 @@
+package au.com.shiftyjelly.pocketcasts.models.db.helper
+
+import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
+
+data class ListenedCategory(
+    val numberOfPodcasts: Int,
+    val totalPlayedTime: Long,
+    val category: String,
+    val mostListenedPodcastId: String,
+    val mostListenedPodcastTintColor: Int,
+) {
+    fun toPodcast() = Podcast(uuid = mostListenedPodcastId, tintColorForLightBg = mostListenedPodcastTintColor)
+}