+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/helper/LongestEpisode.kt
@@ -1,0 +1,19 @@
+package au.com.shiftyjelly.pocketcasts.models.db.helper
+
+import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
+
+data class LongestEpisode(
+    val title: String,
+    val duration: Double,
+    val podcastUuid: String,
+    val podcastTitle: String,
+    val tintColorForLightBg: Int,
+    val tintColorForDarkBg: Int,
+) {
+    fun toPodcast() = Podcast(
+        uuid = podcastUuid,
+        title = podcastTitle,
+        tintColorForLightBg = tintColorForLightBg,
+        tintColorForDarkBg = tintColorForDarkBg,
+    )
+}